+++ packages/engine/Source/Core/clone.js
@@ -14,8 +14,8 @@ function clone(object, deep) {
 
   deep = deep ?? false;
 
-  const Constructor = object.constructor;
-  const result = new Constructor();
+  const ObjectConstructor = object.constructor;
+  const result = new ObjectConstructor();
   for (const propertyName in object) {
     if (object.hasOwnProperty(propertyName)) {
       let value = object[propertyName];