+++ packages/engine/Source/Scene/Primitive.js
@@ -672,8 +672,8 @@ function cloneAttribute(attribute) {
   if (Array.isArray(attribute.values)) {
     clonedValues = attribute.values.slice(0);
   } else {
-    const Constructor = attribute.values.constructor;
-    clonedValues = new Constructor(attribute.values);
+    const ValuesConstructor = attribute.values.constructor;
+    clonedValues = new ValuesConstructor(attribute.values);
   }
   return new GeometryAttribute({
     componentDatatype: attribute.componentDatatype,
@@ -698,8 +698,8 @@ function cloneGeometry(geometry) {
     if (Array.isArray(sourceValues)) {
       indices = sourceValues.slice(0);
     } else {
-      const Constructor = sourceValues.constructor;
-      indices = new Constructor(sourceValues);
+      const IndicesConstructor = sourceValues.constructor;
+      indices = new IndicesConstructor(sourceValues);
     }
   }